analyses/2018_12_ddobre_static_analysis/1-get_script_urls/requirements.txt
appdirs==1.4.3
attrs==18.2.0
Click==7.0
pkg-resources==0.0.0
python-slugify==1.2.6
toml==0.10.0
Unidecode==1.0.23
